pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbeats-tutorial"
version = "0.1.0"
description = "Agentbeats Tutorial"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.5",
    "agentbeats>=1.2.6",
    "google-adk>=1.14.1",
    "google-genai>=1.36.0",
    "litellm>=1.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
agentbeats-run = "agentbeats.run_scenario:main"

[tool.uv]
package = true
dev-dependencies = [
    "mypy>=1.18.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentbeats"]
